docs and changes for docs
