+++ .gitignore
@@ -21,4 +21,5 @@ MoretonVal
 ValData.zip
 Testing
 2020results
+2024results
 testmemory
