--- active_learning/static/image.txt
@@ -1,1 +1,0 @@
-0.683894 0.909856
