+++ counting_boats/__init__.py
@@ -1,0 +1,9 @@
+"""
+This package contains the code for the Counting Boats project. 
+the module boat_utils contains the core functionality for the project,
+while the top-level scripts in this package are for running different parts of the project. 
+
+
+
+Date: 17/09/2024
+"""
