+++ counting_boats/boat_utils/__init__.py
@@ -1,0 +1,8 @@
+"""
+This module contains utility functions for the counting boats project.
+These in general should not be called directly, and are imported by other modules as needed.
+
+Author: Charlie Turner
+
+Date: 17/09/2024
+"""
