+++ counting_boats/boat_utils/config.py
@@ -1,3 +1,11 @@
+"""
+Load and parse the config file for other modules
+
+usage:
+
+    from boat_utils.config import cfg
+"""
+
 import os
 import yaml
