--- data/polygons/moreton_inside.qmd
@@ -1,27 +1,0 @@
-<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
-<qgis version="3.34.2-Prizren">
-  <identifier></identifier>
-  <parentidentifier></parentidentifier>
-  <language></language>
-  <type></type>
-  <title></title>
-  <abstract></abstract>
-  <links/>
-  <dates/>
-  <fees></fees>
-  <encoding></encoding>
-  <crs>
-    <spatialrefsys nativeFormat="Wkt">
-      <wkt></wkt>
-      <proj4></proj4>
-      <srsid>0</srsid>
-      <srid>0</srid>
-      <authid></authid>
-      <description></description>
-      <projectionacronym></projectionacronym>
-      <ellipsoidacronym></ellipsoidacronym>
-      <geographicflag>false</geographicflag>
-    </spatialrefsys>
-  </crs>
-  <extent/>
-</qgis>
