+++ docs/index.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="refresh" content="0; url=./ML_Software/counting_boats.html"/>
+</head>
+</html>
