--- testmemory/orders.csv
@@ -1,5 +1,0 @@
-order_id,aoi,date,order_status,area_coverage,cloud_coverage
-fb4b3a01-c480-43a7-b86f-22e6fa13d450,moreton,2018-07-15,complete,-,-
-1652fbf7-fddd-4a43-8778-84e5d7b50886,moreton,2018-07-16,complete,-,-
-20b160b1-fe2c-4a9e-85fd-fcc481e0e8a6,moreton,2018-07-17,complete,-,-
-1515da24-ad37-4b8a-b235-aeca66fcd016,moreton,2018-07-18,complete,-,-
